[Enhance] Restructure how we keep the components
